Stronger typing

+++ src/data_extraction/get_raw_data.py
@@ -15,7 +15,7 @@ logging.getLogger("ecmwf.datastores").setLevel(logging.WARNING)
 logging.getLogger("cdsapi").setLevel(logging.WARNING)
 
 
-def cut_grib_long_lat(grib_path: str, coordinates: list[int]) -> xr.Dataset:
+def cut_grib_long_lat(grib_path: str, coordinates: list[float]) -> xr.Dataset:
     with xr.open_dataset(
         grib_path,
         engine="cfgrib",
